pyproject.toml
[tool.poetry]
name = "nonebot-plugin-bfchat"
version = "0.2.2"
description = "A battlefield 1/v/2042 chatbot based on nonebot2 framework"
authors = ["Nightsky <050644zf@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_bfchat"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot-plugin-htmlrender = "^0.2.0.1"
nonebot2 = "^2.0.0rc2"
nonebot-adapter-onebot = "^2.2.0"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
